http-triggers/requirements.txt
azure-functions==1.0.0b3
azure-functions-worker==1.0.0b3
grpcio==1.14.2
grpcio-tools==1.14.2
protobuf==3.6.1
six==1.12.0
marshmallow==3.0.0rc1
azure-storage==0.32.0
azure-cosmosdb-table
